basic idea

+++ src/decorators.js
@@ -1,56 +1,46 @@
 
-export function persistable(target) {
-  Object.defineProperty(target.prototype, 'save', {
-    value: save
+export function findById(Repository) {
+  Object.defineProperty(Repository.prototype, 'findById', {
+    value: function findById(id, { $returning = {} } = {}) {
+      const method = 'get'
+      const path = this.buildPath(id)
+      const query = { $returning }
+      return this.sync(method, path, undefined, query)
+    }
   })
 }
 
-function save(instance, { patch = false, $returning = {} } = {}) {
-  const { url, Model } = this
-  const { id } = instance
-  const method = id ? patch ? 'patch' : 'put' : 'post'
-  const path = (method === 'post') ? url : `${url}/${id}`
-  const payload = Model.serialize(instance)
-  const query = { $returning }
-  return this.sync(path, method, payload, query)
-}
-
-export function searchable(target) {
-  Object.defineProperty(target.prototype, 'search', {
-    value: search
+export function save(Repository) {
+  Object.defineProperty(Repository.prototype, 'save', {
+    value: function save(model, { patch = false, $returning = {} } = {}) {
+      const { id } = model
+      const body = this.serialize(model)
+      const method = id != null ? patch ? 'patch' : 'put' : 'post'
+      const path = this.buildPath(id)
+      const query = { $returning }
+      return this.sync(method, path, body, query)
+    }
   })
 }
 
-function search({ $search = '', $returning = {} } = {}) {
-  const method = 'get'
-  const path = this.url
-  const query = { $search, $returning }
-  return this.sync(path, method, undefined, query)
-}
-
-export function listable(target) {
-  Object.defineProperty(target.prototype, 'list', {
-    value: list
+export function list(Repository) {
+  Object.defineProperty(Repository.prototype, 'list', {
+    value: function list(query = {}) {
+      const method = 'get'
+      const path = this.buildPath()
+      return this.sync(method, path, undefined, query)
+    }
   })
 }
 
-function list({ $params = {}, $relations = [] }) {
-  const method = 'get'
-  const path = this.url
-  const query = { $params, $relations }
-  return this.sync(path, method, undefined, query)
-}
-
-export function destroyable(target) {
-  Object.defineProperty(target.prototype, 'destroy', {
-    value: destroy
+export function destroy(Repository) {
+  Object.defineProperty(Repository.prototype, 'destroy', {
+    value: function destroy(model) {
+      const { id } = model
+      if (id == null) return Promise.resolve(this.create({}))
+      const method = 'delete'
+      const path = this.buildPath(id)
+      return this.sync(method, path)
+    }
   })
-}
-
-function destroy(model) {
-  const { url } = this
-  const { id } = model
-  const method = 'delete'
-  const path = `${url}/${id}`
-  return this.sync(path, method)
 }
